+++ changelog.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 #### Unreleased
+- [Fix] Encode the response body based on the `Content-Transfer-Encoding` header.
 
 #### 1.0.0
 - Initial release.